+++ src/config/settings.ts
@@ -90,8 +90,8 @@ const defaultByChain: Record<ChainsIds, BlockchainSettings> = {
     type: resolveBlockchainType(ChainsIds.LINEA) || [BlockchainType.ON_CHAIN],
     transactions: {
       waitForBlockTime: 1000,
-      minGasPrice: 1000000000,
-      maxGasPrice: 100000000000,
+      minGasPrice: 100000000,
+      maxGasPrice: 300000000000,
       mintBalance: {
         warningLimit: '0.01',
         errorLimit: '0.0001',